Split off Path utils
